+++ db-queue-scheduler-core/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/DefaultScheduler.java
@@ -35,8 +35,8 @@ class DefaultScheduler implements Scheduler {
     }
 
     @Override
-    public void schedule(@Nonnull ScheduledTaskSettings scheduledTaskSettings,
-                         @Nonnull ScheduledTask scheduledTask) {
+    public void schedule(@Nonnull ScheduledTask scheduledTask,
+                         @Nonnull ScheduledTaskSettings scheduledTaskSettings) {
         requireNonNull(scheduledTaskSettings, "scheduledTaskSettings");
         requireNonNull(scheduledTask, "scheduledTask");
 
@@ -44,7 +44,7 @@ class DefaultScheduler implements Scheduler {
                 .createExecutionTimeProvider(scheduledTaskSettings.getScheduleSettings());
 
         ScheduledTaskDefinition scheduledTaskDefinition = ScheduledTaskDefinition.builder()
-                .withScheduledTaskIdentity(scheduledTaskSettings.getIdentity())
+                .withEnabled(scheduledTaskSettings.isEnabled())
                 .withMaxExecutionLockInterval(scheduledTaskSettings.getMaxExecutionLockInterval())
                 .withScheduledTask(scheduledTask)
                 .withNextExecutionTimeProvider(executionTimeProvider)
@@ -52,8 +52,7 @@ class DefaultScheduler implements Scheduler {
 
         scheduledTaskManager.schedule(scheduledTaskDefinition);
 
-        log.debug("task scheduled: scheduledTaskSettings={}, scheduledTaskDefinition={}",
-                scheduledTaskSettings, scheduledTaskDefinition);
+        log.info("task scheduled: identity={}, settings={}", scheduledTask.getIdentity(), scheduledTaskSettings);
     }
 
     @Override