+++ db-queue-scheduler-core/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/Scheduler.java
@@ -11,7 +11,7 @@ import java.util.List;
 /**
  * Scheduler manages {@link ScheduledTask}s for periodic execution - configures, registers, starts and pauses.
  *
- * <p>Scheduler guarantees exactly-once task execution that means that any registered tasks is executed exactly once per each
+ * <p>Scheduler guarantees exactly-once task execution that means that any registered tasks are executed exactly once per each
  * scheduled time in spite of any numbers of working application nodes.
  *
  * <p>Scheduler uses RDBMS for persisting registered tasks. Currently, scheduler backed on {@code db-queue} library.
@@ -23,12 +23,12 @@ public interface Scheduler {
     /**
      * Registers a task for periodic executions
      *
-     * @param scheduledTaskSettings settings of the scheduled task
      * @param scheduledTask task for periodic executions
+     * @param scheduledTaskSettings settings of the scheduled task
      * @throws RuntimeException if any scheduled task with the same identity already registered
      */
-    void schedule(@Nonnull ScheduledTaskSettings scheduledTaskSettings,
-                  @Nonnull ScheduledTask scheduledTask);
+    void schedule(@Nonnull ScheduledTask scheduledTask,
+                  @Nonnull ScheduledTaskSettings scheduledTaskSettings);
 
     /**
      * Starts scheduler - makes scheduler available for executing scheduled tasks