+++ db-queue-scheduler-core/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/config/impl/NoopScheduledTaskLifecycleListener.java
@@ -37,6 +37,8 @@ public class NoopScheduledTaskLifecycleListener implements ScheduledTaskLifecycl
 
     /**
      * Get empty listener for task processing lifecycle
+     *
+     * @return flyweight object of empty listener
      */
     public static NoopScheduledTaskLifecycleListener getInstance() {
         return INSTANCE;