+++ db-queue-scheduler-spring/src/test/java/ru/yoomoney/tech/dbqueue/scheduler/SchedulerTest.java
@@ -2,8 +2,9 @@ package ru.yoomoney.tech.dbqueue.scheduler;
 
 import org.junit.jupiter.api.Test;
 import ru.yoomoney.tech.dbqueue.scheduler.config.DatabaseDialect;
+import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTask;
 import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTaskExecutionResult;
-import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTaskIdentity;
+import ru.yoomoney.tech.dbqueue.scheduler.models.SimpleScheduledTask;
 import ru.yoomoney.tech.dbqueue.scheduler.settings.ScheduleSettings;
 import ru.yoomoney.tech.dbqueue.scheduler.settings.ScheduledTaskSettings;
 
@@ -31,19 +32,22 @@ public class SchedulerTest extends BaseTest {
                 .withTransactionOperations(transactionTemplate)
                 .configure();
         AtomicBoolean executed = new AtomicBoolean(false);
-
-        // when
-        scheduler.start();
-        scheduler.schedule(
-                ScheduledTaskSettings.builder()
-                        .withIdentity(ScheduledTaskIdentity.of("scheduled-task" + uniqueCounter.incrementAndGet()))
-                        .withMaxExecutionLockInterval(Duration.ofHours(1L))
-                        .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(0L)))
-                        .build(),
+        ScheduledTask scheduledTask = SimpleScheduledTask.create(
+                "scheduled-task" + uniqueCounter.incrementAndGet(),
                 () -> {
                     executed.set(true);
                     return ScheduledTaskExecutionResult.success();
                 }
+        );
+
+        // when
+        scheduler.start();
+        scheduler.schedule(
+                scheduledTask,
+                ScheduledTaskSettings.builder()
+                        .withMaxExecutionLockInterval(Duration.ofHours(1L))
+                        .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(0L)))
+                        .build()
         );
 
         // then
@@ -65,29 +69,31 @@ public class SchedulerTest extends BaseTest {
                 .withJdbcOperations(jdbcTemplate)
                 .withTransactionOperations(transactionTemplate)
                 .configure();
-        String taskName = "scheduled-task" + uniqueCounter.incrementAndGet();
+        ScheduledTask scheduledTask = SimpleScheduledTask.create(
+                "scheduled-task" + uniqueCounter.incrementAndGet(),
+                ScheduledTaskExecutionResult::success
+        );
 
         // when
         scheduler1.schedule(
+                scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withIdentity(ScheduledTaskIdentity.of(taskName))
                         .withMaxExecutionLockInterval(Duration.ofHours(1L))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(1L)))
-                        .build(),
-                ScheduledTaskExecutionResult::success
+                        .build()
         );
         scheduler2.schedule(
+                scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withIdentity(ScheduledTaskIdentity.of(taskName))
                         .withMaxExecutionLockInterval(Duration.ofHours(1L))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(1L)))
-                        .build(),
-                ScheduledTaskExecutionResult::success
+                        .build()
         );
 
         // then
         assertThat(
-                jdbcTemplate.queryForObject("select count(1) from scheduled_tasks where queue_name = ?", Long.class, taskName),
+                jdbcTemplate.queryForObject("select count(1) from scheduled_tasks where queue_name = ?", Long.class,
+                        scheduledTask.getIdentity().asString()),
                 equalTo(1L)
         );
     }
@@ -101,25 +107,26 @@ public class SchedulerTest extends BaseTest {
                 .withJdbcOperations(jdbcTemplate)
                 .withTransactionOperations(transactionTemplate)
                 .configure();
-        String taskName = "scheduled-task" + uniqueCounter.incrementAndGet();
+        ScheduledTask scheduledTask = SimpleScheduledTask.create(
+                "scheduled-task" + uniqueCounter.incrementAndGet(),
+                ScheduledTaskExecutionResult::success
+        );
 
         // when
         scheduler.schedule(
+                scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withIdentity(ScheduledTaskIdentity.of(taskName))
                         .withMaxExecutionLockInterval(Duration.ofHours(1L))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(1L)))
-                        .build(),
-                ScheduledTaskExecutionResult::success
+                        .build()
         );
 
         // then
         assertThrows(RuntimeException.class, () -> scheduler.schedule(
+                scheduledTask,
                 ScheduledTaskSettings.builder()
-                        .withIdentity(ScheduledTaskIdentity.of(taskName))
                         .withMaxExecutionLockInterval(Duration.ofHours(1L))
                         .withScheduleSettings(ScheduleSettings.fixedDelay(Duration.ofSeconds(1L)))
-                        .build(),
-                ScheduledTaskExecutionResult::success));
+                        .build()));
     }
 }